Add files via upload
Changed pump on/off times. Removed duplicate code for DateTime now from setup as it is only needed for the logging section in the loop.  Commented out line 654 for troubleshooting wrong values in intervalUnixTime.  Fan is still coming on when the measured lux is greater than 0. Don't understand why yet.